修改分值/攻击/生命数值的基础倍率为10

+++ Assets/Script/Effect/Leader/SPToken/AggregatorEffect.cs
@@ -3,7 +3,7 @@
 public class AggregatorEffect : Effect
 {
     private string _field = "";
-    private float _increment = 0;
+    private int _increment = 0;
     private string _seriesId = "";
 
     public override string Name() => "集结者";
